requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose==3.3.0
cryptography==44.0.3
httpx==0.27.0
supabase==2.15.0
anthropic==0.50.0
slowapi==0.1.9
bcrypt==4.2.1
pdfplumber==0.11.4
openpyxl==3.1.5
python-docx==1.1.2
python-multipart==0.0.20
google-auth==2.40.3
google-auth-oauthlib==1.2.1
google-api-python-client==2.169.0
beautifulsoup4==4.13.4
